updated handlers returning null

+++ backend/pkg/server/handlers/accounts/unlink_account.go
@@ -2,7 +2,6 @@ package handlers
 
 import (
 	"context"
-	"encoding/json"
 	"github.com/jackc/pgx/v5/pgxpool"
 	"net/http"
 )
@@ -12,10 +11,7 @@ func UnlinkAccountHandler(pool *pgxpool.Pool) http.HandlerFunc {
 		provider := r.PathValue("provider")
 		providerAccountId := r.PathValue("providerAccountId")
 
-		sql := `
-delete from accounts
-       where provider=$1 and
-             provider_account_id=$2`
+		sql := "delete from accounts where provider=$1 and provider_account_id=$2"
 
 		commandTag, err := pool.Exec(context.Background(), sql, provider, providerAccountId)
 		if err != nil {
@@ -28,7 +24,6 @@ delete from accounts
 			return
 		}
 
-		nilResponse, _ := json.Marshal(nil)
-		w.Write(nilResponse)
+		_, _ = w.Write([]byte("null"))
 	}
 }
